fix: use BigInteger instead of Long in Balance & Token.

+++ solana-common/src/main/kotlin/com/rarible/protocol/solana/common/converter/BalanceConverter.kt
@@ -6,7 +6,7 @@ import com.rarible.solana.protocol.dto.BalanceDto
 object BalanceConverter {
     fun convert(balance: Balance): BalanceDto = BalanceDto(
         account = balance.account,
-        value = balance.value.toBigInteger(),
+        value = balance.value,
         createdAt = balance.createdAt,
         updatedAt = balance.updatedAt
     )
